solve merges coming to an incorrect consensus
At the end of the script, we would check if games that were not in the original manifest get removed out. Since equality checks are a bit flawed by nature, if the date changed was different then this check would fail to find it in the list, even though it was there. So we double-check, by grabbing the elements, and making sure they exist in both, just in-case.

This also solves another bug, where elements could be duplicated because of this. Early on, we check if there are changes, and that they are newer then what's in the manifest, if not we move on. However in some cases, edits can have multiple different changes, or be delayed compare to other edits that are also coming in. So even if the changes aren't newer, we need to make sure they are marked as edited, so they're not deleted at the end of the script too.